requirements.txt
crispy-bootstrap4==2024.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
